Update README.md
The goal of this project is to develop an online student course evaluation system (SCES) to replace the paper-based evaluation system currently used at Faculty of Computing. The SCES was developed using iterative enhancement software development model to accommodate the changing and new requirements as the system evolved. Key features including the development of evaluation forms and analysis of the submitted forms for report statistics are successfully implemented. Functionalities of SCES are designed and implemented in such a way that they are accessible only to the registered user as per their system privileges. Each user in SCES is given a specific role. Students can fill the evaluation forms and submit them. Administrators can create/modify/delete user accounts, courses, surveys, questions and comments for evaluation forms to be completed by the students. The administrator can view the evaluation results and send the result to staff. The system was tested for functionality and browser efficiency. 

	Database Connection
The system is connected to the database from this code. The die condition can be used if any error is occurred in the connection. Database can be connected using Mysql through Apache, Xampp program, this will connect the frontend and the backend of the software program.

Login Module
The system shall allow the administrator to create, edit, and delete evaluation form and allow him to add survey for a particular course.

ii.The system shall allow the administrator to upload students email and add lecturer name and his email.

iii.The system shall allow the administrator to add, edit and delete courses.

iv.The system shall allow the administrator to analyze survey result.

v.The system shall  allow the administrator to send survey result to Deans, HOD, lecturer, and student.

vi.The system shall allow the administrator to send the evaluation form to the student for a particular course.

vii.The system shall allow the student to conduct course evaluation.

viii.The system shall allow the administrator to design an evaluation form consisting of multiple choice, slider and 
open-ended questions asking for students' opinions about various aspects of a course.

ix.A student should be able to go back and forth between the pages of an evaluation for a particular course. Once the student submits the evaluation form for a course, then the student/administrator can’t change the information entered. Hence, students can’t evaluate that course again.

x.The system shall only save the responses of the student for the questionnaire and No personal identification of the student should be saved.

The following steps is employed so as to meet the objectives of this work:
A database is  designed capable of storing users’ information and Evaluation Results
MySQL is used as a database platform for the system implementation.
PHP is used as a server-side scripting language for this project.
JAVASCRIPT is an object-oriented scripting language that enables programmatic access to computational objects within the host environment.
HTML and CSS is used to a design graphical user interface for the system.
IDE used in this project is Visual Studio Code.
Web Server in this project is WampServer.